+++ src/main/kotlin/com/rtomyj/skc/browse/card/CardBrowseService.kt
@@ -5,8 +5,8 @@ import com.rtomyj.skc.browse.card.model.Card
 import com.rtomyj.skc.browse.card.model.CardBrowseCriteria
 import com.rtomyj.skc.browse.card.model.CardBrowseResults
 import com.rtomyj.skc.browse.card.model.MonsterAssociation
-import com.rtomyj.skc.enums.MonsterAssociationExpression
-import com.rtomyj.skc.enums.MonsterAssociationType
+import com.rtomyj.skc.util.enumeration.MonsterAssociationExpression
+import com.rtomyj.skc.util.enumeration.MonsterAssociationType
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Qualifier
