+++ src/main/kotlin/com/rtomyj/skc/browse/card/dao/CardBrowseJDBCDao.kt
@@ -5,9 +5,9 @@ import com.rtomyj.skc.browse.card.model.Card
 import com.rtomyj.skc.browse.card.model.CardBrowseCriteria
 import com.rtomyj.skc.browse.card.model.CardBrowseResults
 import com.rtomyj.skc.browse.card.model.MonsterAssociation
-import com.rtomyj.skc.constant.DBQueryConstants
-import com.rtomyj.skc.enums.MonsterAssociationType
-import com.rtomyj.skc.enums.table.definitions.BrowseQueryDefinition
+import com.rtomyj.skc.util.constant.DBQueryConstants
+import com.rtomyj.skc.util.enumeration.MonsterAssociationType
+import com.rtomyj.skc.util.enumeration.table.definitions.BrowseQueryDefinition
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Qualifier
