+++ src/main/kotlin/com/rtomyj/skc/browse/product/model/ProductContent.kt
@@ -1,7 +1,7 @@
 package com.rtomyj.skc.browse.product.model
 
 import com.fasterxml.jackson.annotation.JsonInclude
-import com.rtomyj.skc.constant.SwaggerConstants
+import com.rtomyj.skc.util.constant.SwaggerConstants
 import com.rtomyj.skc.browse.card.model.Card
 import io.swagger.v3.oas.annotations.media.Schema
 import org.springframework.hateoas.RepresentationModel
