+++ src/main/kotlin/com/rtomyj/skc/browse/product/model/Products.kt
@@ -1,9 +1,9 @@
 package com.rtomyj.skc.browse.product.model
 
 import com.fasterxml.jackson.annotation.JsonInclude
-import com.rtomyj.skc.constant.SwaggerConstants
+import com.rtomyj.skc.util.constant.SwaggerConstants
 import com.rtomyj.skc.browse.product.ProductBrowseController
-import com.rtomyj.skc.enums.ProductType
+import com.rtomyj.skc.util.enumeration.ProductType
 import com.rtomyj.skc.util.HateoasLinks
 import io.swagger.v3.oas.annotations.media.Schema
 import org.springframework.hateoas.RepresentationModel
