+++ src/main/kotlin/com/rtomyj/skc/find/banlist/dao/BanListDao.kt
@@ -1,6 +1,6 @@
 package com.rtomyj.skc.banlist.dao
 
-import com.rtomyj.skc.enums.BanListCardStatus
+import com.rtomyj.skc.util.enumeration.BanListCardStatus
 import com.rtomyj.skc.exception.YgoException
 import com.rtomyj.skc.banlist.model.BanListDates
 import com.rtomyj.skc.banlist.model.CardBanListStatus
