+++ src/main/kotlin/com/rtomyj/skc/find/banlist/dao/BanListHibernateDao.kt
@@ -1,7 +1,7 @@
 package com.rtomyj.skc.banlist.dao
 
-import com.rtomyj.skc.constant.ErrConstants
-import com.rtomyj.skc.enums.BanListCardStatus
+import com.rtomyj.skc.util.constant.ErrConstants
+import com.rtomyj.skc.util.enumeration.BanListCardStatus
 import com.rtomyj.skc.exception.ErrorType
 import com.rtomyj.skc.exception.YgoException
 import com.rtomyj.skc.banlist.model.BanListDate
