+++ src/main/kotlin/com/rtomyj/skc/find/banlist/model/BanListDate.kt
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonInclude
 import com.rtomyj.skc.banlist.controller.BanListDiffController
 import com.rtomyj.skc.banlist.controller.BannedCardsController
 import com.rtomyj.skc.config.DateConfig
-import com.rtomyj.skc.constant.SwaggerConstants
+import com.rtomyj.skc.util.constant.SwaggerConstants
 import com.rtomyj.skc.util.HateoasLinks
 import io.swagger.v3.oas.annotations.media.Schema
 import org.springframework.hateoas.RepresentationModel
