+++ src/main/kotlin/com/rtomyj/skc/find/card/CardController.kt
@@ -1,7 +1,7 @@
 package com.rtomyj.skc.find.card
 
-import com.rtomyj.skc.constant.SKCRegex
-import com.rtomyj.skc.constant.SwaggerConstants
+import com.rtomyj.skc.util.constant.SKCRegex
+import com.rtomyj.skc.util.constant.SwaggerConstants
 import com.rtomyj.skc.util.YgoApiBaseController
 import com.rtomyj.skc.exception.YgoError
 import com.rtomyj.skc.exception.YgoException
