+++ src/main/kotlin/com/rtomyj/skc/search/card/dao/CardSearchJDBCDao.kt
@@ -2,7 +2,7 @@ package com.rtomyj.skc.search.card.dao
 
 import com.rtomyj.skc.banlist.model.CardBanListStatus
 import com.rtomyj.skc.browse.card.model.Card
-import com.rtomyj.skc.constant.DBQueryConstants
+import com.rtomyj.skc.util.constant.DBQueryConstants
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Qualifier
