+++ src/main/kotlin/com/rtomyj/skc/stats/dao/StatsJDBCDao.kt
@@ -1,6 +1,6 @@
 package com.rtomyj.skc.stats.dao
 
-import com.rtomyj.skc.constant.DBQueryConstants
+import com.rtomyj.skc.util.constant.DBQueryConstants
 import com.rtomyj.skc.stats.model.DatabaseStats
 import com.rtomyj.skc.stats.model.MonsterTypeStats
 import org.slf4j.LoggerFactory
