+++ src/main/kotlin/com/rtomyj/skc/util/constant/AppConstants.kt
@@ -1,4 +1,4 @@
-package com.rtomyj.skc.constant
+package com.rtomyj.skc.util.constant
 
 object AppConstants {
 	var APP_VERSION: String = AppConstants::class.java.getPackage().implementationVersion ?: "LOCAL"
