+++ src/main/kotlin/com/rtomyj/skc/util/constant/LogConstants.kt
@@ -1,4 +1,4 @@
-package com.rtomyj.skc.constant
+package com.rtomyj.skc.util.constant
 
 object LogConstants {
 	const val EXCEPTION_PROVIDER_LOG = "Exception occurred w/ message: {}, with ErrorType: {}, responding with: {}"
