+++ src/main/kotlin/com/rtomyj/skc/util/enumeration/MonsterAssociationType.kt
@@ -1,4 +1,4 @@
-package com.rtomyj.skc.enums
+package com.rtomyj.skc.util.enumeration
 
 enum class MonsterAssociationType(
     private val assocationType: String
