+++ src/main/kotlin/com/rtomyj/skc/util/enumeration/Rarity.kt
@@ -1,4 +1,4 @@
-package com.rtomyj.skc.enums
+package com.rtomyj.skc.util.enumeration
 
 enum class Rarity(private val rarityName: String) {
     C("Common"),
