+++ src/main/kotlin/com/rtomyj/skc/util/enumeration/table/definitions/BrowseQueryDefinition.kt
@@ -1,4 +1,4 @@
-package com.rtomyj.skc.enums.table.definitions
+package com.rtomyj.skc.util.enumeration.table.definitions
 
 enum class BrowseQueryDefinition(
     private val columnName: String
