+++ src/test/kotlin/com/rtomyj/skc/find/card/CardControllerTest.kt
@@ -1,7 +1,7 @@
 package com.rtomyj.skc.find.card
 
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
-import com.rtomyj.skc.constant.ErrConstants
+import com.rtomyj.skc.util.constant.ErrConstants
 import com.rtomyj.skc.constant.TestConstants
 import com.rtomyj.skc.exception.ErrorType
 import com.rtomyj.skc.exception.YgoException
